Separate cleanup logic from extensions. Cleanup tests without GPU packages
