+++ ext/CodeDiffsAMDGPU.jl
@@ -46,25 +46,8 @@ function highlight_using_pygments(io::IO, str::AbstractString, lexer)
     end
 end
 
-
-function CodeDiffs.cleanup_code(::Val{:rocm_typed}, c, dbinfo, cleanup_opts)
-    if get(cleanup_opts, :expand_llvmcall, true)
-        c = CodeDiffs.cleanup_inline_llvmcall_modules(c)
-    end
-    return c
-end
-
-CodeDiffs.cleanup_code(::Val{:rocm_llvm}, c, dbinfo, cleanup_opts) = CodeDiffs.replace_llvm_module_name(c)
-
-CodeDiffs.cleanup_code(::Val{:gcn}, c, dbinfo, cleanup_opts) =
-    CodeDiffs.cleanup_code(Val{:rocm_native}(), c, dbinfo, cleanup_opts)
-
-function CodeDiffs.cleanup_code(::Val{:rocm_native}, c, dbinfo, cleanup_opts)
-    c = CodeDiffs.replace_llvm_module_name(c)
-    # Remove the hundreds of '.ident "clang version ..."' lines
-    c = replace(c, r"\t\.ident\t.+\n" => "")
-    # We could also remove everything after the "; -- End function", if anybody wants it
-    return c
-end
+CodeDiffs.cleanup_code(::Val{:rocm_typed},  c, dbinfo, cleanup_opts) = CodeDiffs.cleanup_code(Val{:gpu_typed}(), c, dbinfo, cleanup_opts)
+CodeDiffs.cleanup_code(::Val{:rocm_llvm},   c, dbinfo, cleanup_opts) = CodeDiffs.cleanup_code(Val{:gpu_llvm}(), c, dbinfo, cleanup_opts)
+CodeDiffs.cleanup_code(::Val{:rocm_native}, c, dbinfo, cleanup_opts) = CodeDiffs.cleanup_code(Val{:gcn}(), c, dbinfo, cleanup_opts)
 
 end
