+++ src/CodeDiffs.jl
@@ -20,10 +20,12 @@ const ANSI_REGEX = r"(?>\x1B\[[0-?]*[ -/]*[@-~])+"
 const OhMYREPL_PKG_ID = Base.PkgId(Base.UUID("5fb14364-9ced-5910-84b2-373655c76a03"), "OhMyREPL")
 const Revise_PKG_ID = Base.PkgId(Base.UUID("295af30f-e4ad-537b-8983-00126c2a3abe"), "Revise")
 
+include("cleanup/cleanup.jl")
+using .Cleanup
+
 include("CodeDiff.jl")
 include("get_code.jl")
 include("highlighting.jl")
-include("cleanup.jl")
 include("compare.jl")
 include("display.jl")
