+++ src/cleanup/agx.jl
@@ -1,0 +1,5 @@
+
+function cleanup_code(::Val{:agx}, c, dbinfo, cleanup_opts)
+    c = replace_llvm_module_name(c)
+    return c
+end
