+++ src/cleanup/ptx.jl
@@ -1,0 +1,206 @@
+
+# The PTX output of LLVM's NVPTX backend has a comment with the mangled function name
+const PTX_FUNC_NAME_COMMENT_REGEX = r"// -- Begin function "m * MANGLED_NAME_REGEX * r"$"m
+
+# Same with SASS but the name is surrounded by a lot of dashes
+const SASS_FUNC_NAME_COMMENT_REGEX = r"//-{5,} .text."m * MANGLED_NAME_REGEX * r" -{5,}$"m
+
+
+function cleanup_code(::Val{:ptx}, c, dbinfo, cleanup_opts)
+    c = clean_function_name(PTX_FUNC_NAME_COMMENT_REGEX, c)
+    c = cleanup_external_functions(c)
+
+    extra_patterns = []
+    if get(cleanup_opts, :demangle, true)
+        # Demangle only names which aren't the function's name
+        push!(extra_patterns, demangle_all())
+    end
+    if get(cleanup_opts, :indent_calls, true)
+        push!(extra_patterns, indent_ptx_function_calls())
+    end
+    if !get(cleanup_opts, :keep_loop_comments, false)
+        # Loop comments are interesting, but can be quite annoying with complex loop structures
+        push!(extra_patterns, r"\n\s+\/\/\s+(=>\s*)?(Parent|This|This Inner|Child|in) Loop[^\n]+" => "")
+    end
+    if !get(cleanup_opts, :keep_block_comments, false)
+        # Block comments are placed after labels, or at the beginning of a line
+        # They look like "some_label: // %blabla" or "// %blabla".
+        push!(extra_patterns, r"^([a-zA-Z0-9_$%]+:)?(?:\s*)\/\/ %.+"m => s"\1")
+    end
+    if (align_preds = get(cleanup_opts, :align_preds, 8); align_preds > 0)
+        # Align guard predicates such that the beginning of the instruction is at the
+        # `align_preds`-th column.
+        push!(extra_patterns, align_instruction_predicates(align_preds))
+    end
+
+    return replace(c,
+        # Parameter names are always local to a function, so this should never create ambiguous code.
+        trim_parameter_names(),
+        # Generic match to cleanup external functions as well
+        llvm_module_name_patterns()...,
+        # Remove the 3 extra header lines
+        r"//\s//\sGenerated by LLVM NVPTX Back-End\s//\s\s" => "",
+        # I don't know what `callseq` indicates in comments, but the numbers after change at every call
+        r" // callseq .+$"m => "",
+        # Remove 'inline asm' comments (remove the newline as well)
+        r"^\s+// (begin|end) inline asm\n"m => "",
+        # Remove the "end function" comment
+        r"^\s+// -- End function\n"m => "",
+        # Remove empty lines
+        r"\n{2,}" => "\n",
+        extra_patterns...
+    )
+end
+
+
+function cleanup_code(::Val{:sass}, c, dbinfo, cleanup_opts)
+    # SASS problems:
+    # Registers seem to be assigned randomly, changing from one call to another with the
+    # same input, as well as some immediate values (maybe related to the functions order in PTX?).
+    # Some instructions (only `MOV` from what I saw) might be ordered differently, or even
+    # in different numbers (max of what I could see is 1, but still surprising).
+    c = replace_llvm_module_name(c)
+    c = clean_function_name(SASS_FUNC_NAME_COMMENT_REGEX, c)
+
+    extra_patterns = []
+    if get(cleanup_opts, :demangle, true)
+        # Demangle only names which aren't the function's name
+        push!(extra_patterns, demangle_all())
+    end
+    if !dbinfo
+        push!(extra_patterns, r"; Location .+\n" => "")  # Remove location comments
+    end
+
+    return replace(c, extra_patterns...)
+end
+
+
+function find_main_ptx_function(c)
+    # Matches any declaration of the form ".a1 .a2 some_name ( ..." and extracts ".a1 .a2" in group
+    # "attrs" and "some_name" in group "name".
+    # Since there are multiple ways to mark a function as the entry point, with ".entry" from the
+    # PTX itself, or with ".func" (without ".external" in its attributes), we must parse through the
+    # code until we find an acceptable candidate.
+    func_name_regex = r"(?<attrs>(?:\.\w+\s*)+)\s+(?<name>[a-zA-Z0-9_$%]+)\s*\("m
+
+    func_name = nothing
+    for m in eachmatch(func_name_regex, c)
+        is_entry  = occursin(".entry", m[:attrs])
+        is_entry |= !occursin(".external", m[:attrs]) && occursin(".visible", m[:attrs])  # => ".visible" implies that it is visible from outside the module
+        !is_entry && continue
+        func_name = String(m[:name])
+        break
+    end
+
+    isnothing(func_name) && return nothing
+
+    # The name may still be decorated with the LLVM module number, etc...
+    m = match(function_unique_gen_name_regex(), func_name)
+    return String(@something m[1] m[2] func_name)
+end
+
+
+function cleanup_external_functions(c)
+    # In PTX code there can be external functions defined with the actual kernel function.
+    # Those can have random positions, and their parameters can have names using the
+    # `globalUniqueGeneratedNames` value.
+    # To make the output consistent each time, we must filter the parameter names and order
+    # the function definitions in a deterministic way.
+
+    # Match external function definitions by following the PTX syntax:
+    # https://docs.nvidia.com/cuda/parallel-thread-execution/index.html?highlight=extern#linking-directives-extern
+    # https://docs.nvidia.com/cuda/parallel-thread-execution/index.html?highlight=extern#kernel-and-function-directives-func
+    re_extern_func = r"\.extern \.func\s+"s
+    re_attribute_list = r"(?:\.attribute\(.+?\)\s+)?"s
+    re_retval = r"(?:\(.+\)\s+)?"s
+    re_func_name = r"(\w+)\s+"s  # Capture the function name
+    re_param_list = r"\([^\)]+\)"s
+    re_noreturn = r"(?:\s+\.noreturn)?"s
+    re_func_end = r"\s*;"s       # Since it is an external function it has no body
+    re_extern_func_def = re_extern_func * re_attribute_list * re_retval * re_func_name * re_param_list * re_noreturn * re_func_end
+
+    external_funcs = String[]
+    for external_func in eachmatch(re_extern_func_def, c)
+        func_def  = external_func.match
+        func_name = external_func[1]
+
+        param_idx = 0
+        function rename_bad_param(_)
+            new_name = func_name * "_param_" * string(param_idx)
+            param_idx += 1
+            return new_name
+        end
+
+        func_def = replace(func_def, func_name * r"_\d+_param" => rename_bad_param)
+        push!(external_funcs, func_def)
+    end
+
+    isempty(external_funcs) && return c
+    sort!(external_funcs)  # Basic attempt at ordering external functions, this might not be enough
+
+    func_idx = 0
+    function next_func_def(_)
+        func_idx += 1
+        return external_funcs[func_idx]
+    end
+
+    # Parse the code again but this time replace the external functions with the sorted ones
+    return replace(c, re_extern_func_def => next_func_def)
+end
+
+
+function trim_parameter_names()
+    # Matches "blabla_param_1" and replaces it by "param_1"
+    return r"\b[a-zA-Z0-9_$%]+_(param_\d+)\b" => s"\1"
+end
+
+
+function indent_ptx_function_calls()
+    # Matches `call func, (param_1, param_2...);`
+    # Groups:
+    #  - 1: call instruction options
+    #  - 2: `func`
+    #  - 3: parameters list
+    ptx_call_regex = r"call(.*)\s+(.+),\s+\(([^;]+)\n\s+\);"
+
+    function indent_call_params(call_inst)
+        m = match(ptx_call_regex, call_inst)
+        isnothing(m) && return call_inst
+        indent = " "^8
+        params = replace(m[3], "\n" => "\n" * indent)
+        return "call" * m[1] * m[2] * ", (" * params * "\n" * indent * ");"
+    end
+
+    return ptx_call_regex => indent_call_params
+end
+
+
+function align_instruction_predicates(column)
+    # The regex extracts from "   @p12 any_instruction" or "    @!p12 any_instruction" the "@p12" or
+    # "@!p12" predicate, without the surrounding spaces. The "any_instruction" part is in group 2.
+    pred_regex = r"^\s*(@!?[^\s]+)\s*([^\s]+);?"m
+
+    function align_to_nth_column(c)
+        m = match(pred_regex, c)
+        isnothing(m) && return c
+        if !endswith(m.match, ';')
+            # If the instruction has operands, then we need to add spaces after the mnemonic such
+            # that they stay in their initial column, so that they stay aligned with the other
+            # instructions' operands.
+            # e.g. "    @p1 bra    $L2;"  =>  " @p1 bra    $L2;"  =>  " @p1 bra        $L2;"
+            operands_column = column * 2  # By default, operands are placed after the 16th column
+            pred_length     = length(m[1]) + 1  # +1 because of the space afterward
+            mnemonic_length = length(m[2])
+            if mnemonic_length + column > operands_column
+                extra_spaces = ""  # the mnemonic is too long
+            else
+                extra_spaces = " "^min(pred_length, column)
+            end
+        else
+            extra_spaces = ""
+        end
+        return lpad(m[1], column - 1) * " " * m[2] * extra_spaces
+    end
+
+    return pred_regex => align_to_nth_column
+end
