+++ test/samples/arithmetic_func.ptx
@@ -1,0 +1,57 @@
+//
+// Generated by LLVM NVPTX Back-End
+//
+
+.version 8.5
+.target sm_80
+.address_size 64
+
+	// .globl	julia_tid_to_blk_index_warp_aware_thread_side_8760 // -- Begin function julia_tid_to_blk_index_warp_aware_thread_side_8760
+                                        // @julia_tid_to_blk_index_warp_aware_thread_side_8760
+.visible .func julia_tid_to_blk_index_warp_aware_thread_side_8760(
+	.param .b64 julia_tid_to_blk_index_warp_aware_thread_side_8760_param_0,
+	.param .b64 julia_tid_to_blk_index_warp_aware_thread_side_8760_param_1,
+	.param .b64 julia_tid_to_blk_index_warp_aware_thread_side_8760_param_2
+)
+{
+	.reg .pred 	%p<4>;
+	.reg .b32 	%r<16>;
+	.reg .b64 	%rd<15>;
+
+// %bb.0:                               // %top
+	ld.param.u64 	%rd5, [julia_tid_to_blk_index_warp_aware_thread_side_8760_param_2];
+	ld.param.u64 	%rd4, [julia_tid_to_blk_index_warp_aware_thread_side_8760_param_0];
+	ld.param.u64 	%rd6, [julia_tid_to_blk_index_warp_aware_thread_side_8760_param_1];
+	ld.u32 	%r15, [%rd5+12];
+	shl.b32 	%r7, %r15, 5;
+	ld.u64 	%rd7, [%rd6+8];
+	cvt.u32.u64 	%r2, %rd7;
+	add.s64 	%rd8, %rd7, -1;
+	cvt.u64.u32 	%rd9, %r7;
+	mul.lo.s64 	%rd10, %rd8, %rd9;
+	ld.u64 	%rd11, [%rd6];
+	add.s64 	%rd14, %rd10, %rd11;
+	ld.v2.u32 	{%r3, %r4}, [%rd5];
+	sub.s32 	%r8, %r3, %r4;
+	setp.ge.u32 	%p1, %r8, %r2;
+	@%p1 bra 	$L__BB0_2;
+// %bb.1:                               // %L17
+	add.s32 	%r15, %r15, 1;
+	not.b32 	%r9, %r3;
+	add.s32 	%r10, %r9, %r2;
+	add.s32 	%r11, %r10, %r4;
+	mul.wide.u32 	%rd12, %r11, 32;
+	add.s64 	%rd14, %rd12, %rd14;
+$L__BB0_2:                              // %L27
+	ld.u32 	%r12, [%rd5+8];
+	setp.ne.s32 	%p2, %r12, 0;
+	setp.eq.s32 	%p3, %r3, %r2;
+	selp.b32 	%r13, %r12, 32, %p3;
+	selp.b32 	%r14, %r13, 32, %p2;
+	ld.u64 	%rd13, [%rd5+16];
+	st.u64 	[%rd4], %rd14;
+	st.u64 	[%rd4+8], %rd13;
+	st.v2.u32 	[%rd4+16], {%r15, %r14};
+	ret;
+                                        // -- End function
+}
