+++ test/samples/reverse_kernel.ptx
@@ -1,0 +1,101 @@
+//
+// Generated by LLVM NVPTX Back-End
+//
+
+.version 8.5
+.target sm_80
+.address_size 64
+
+	// .globl	_Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE // -- Begin function _Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE
+.extern .func julia_throw_boundserror_2464
+(
+	.param .align 8 .b8 julia_throw_boundserror_2464_param_0[16]
+)
+;
+// shmem has been demoted
+                                        // @_Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE
+.visible .entry _Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE(
+	.param .align 8 .b8 _Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE_param_0[16],
+	.param .align 8 .b8 _Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE_param_1[16],
+	.param .align 8 .b8 _Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE_param_2[32]
+)
+.maxntid 128, 1, 1
+{
+	.reg .pred 	%p<4>;
+	.reg .b32 	%r<5>;
+	.reg .b64 	%rd<24>;
+	// demoted variable
+	.shared .align 32 .b8 shmem[16];
+// %bb.0:                               // %conversion
+	ld.param.u64 	%rd11, [_Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE_param_1];
+	mov.u32 	%r3, %ctaid.x;
+	mov.u32 	%r1, %tid.x;
+	add.s32 	%r4, %r1, 1;
+	cvt.u64.u32 	%rd12, %r4;
+	mul.wide.u32 	%rd2, %r3, 128;
+	add.s64 	%rd3, %rd2, %rd12;
+	setp.gt.s64 	%p1, %rd3, %rd11;
+	@%p1 bra 	$L__BB0_6;
+// %bb.1:                               // %L104
+	ld.param.u64 	%rd10, [_Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE_param_2+24];
+	ld.param.u32 	%r2, [_Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE_param_0+8];
+	ld.param.u64 	%rd6, [_Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE_param_0];
+	setp.le.s64 	%p2, %rd3, %rd10;
+	@%p2 bra 	$L__BB0_3;
+	bra.uni 	$L__BB0_2;
+$L__BB0_3:                              // %L246
+	ld.param.u64 	%rd7, [_Z14gpu_reverse_ka16CompilerMetadataI11DynamicSize12DynamicCheckv16CartesianIndicesILi1E5TupleI5OneToI5Int64EEE7NDRangeILi1ES0_10StaticSizeI6_128__ES8_vEE13CuDeviceArrayIS5_Li1ELi1EE_param_2];
+	cvt.u64.u32 	%rd13, %r1;
+	add.s64 	%rd14, %rd2, %rd13;
+	add.s64 	%rd4, %rd14, -1;
+	shl.b64 	%rd15, %rd4, 3;
+	add.s64 	%rd5, %rd7, %rd15;
+	ld.global.u64 	%rd16, [%rd5+8];
+	shl.b64 	%rd17, %rd3, 3;
+	mov.u64 	%rd18, shmem;
+	sub.s64 	%rd19, %rd18, %rd17;
+	st.shared.u64 	[%rd19+16], %rd16;
+	bar.sync 	0;
+	setp.lt.u64 	%p3, %rd3, 3;
+	@%p3 bra 	$L__BB0_5;
+	bra.uni 	$L__BB0_4;
+$L__BB0_5:                              // %L274
+	add.s64 	%rd22, %rd18, %rd15;
+	ld.shared.u64 	%rd23, [%rd22+8];
+	st.global.u64 	[%rd5+8], %rd23;
+$L__BB0_6:                              // %L281
+	ret;
+$L__BB0_2:                              // %L226
+	{ // callseq 16, 0
+	.reg .b32 temp_param_reg;
+	.param .align 8 .b8 param0[16];
+	st.param.b64 	[param0+0], %rd6;
+	st.param.b32 	[param0+8], %r2;
+	call.uni 
+	julia_throw_boundserror_2464, 
+	(
+	param0
+	);
+	} // callseq 16
+	trap;
+	// begin inline asm
+	exit;
+	// end inline asm
+$L__BB0_4:                              // %L256
+	{ // callseq 15, 0
+	.reg .b32 temp_param_reg;
+	.param .align 8 .b8 param0[16];
+	st.param.b64 	[param0+0], %rd6;
+	st.param.b32 	[param0+8], %r2;
+	call.uni 
+	julia_throw_boundserror_2464, 
+	(
+	param0
+	);
+	} // callseq 15
+	trap;
+	// begin inline asm
+	exit;
+	// end inline asm
+                                        // -- End function
+}
